Low: Core: Another cast when printing uint64_t
Mercurial revision: 8f88e8d0741b0c6fc7e4faa07ce51b2705bc73e2

+++ lib/crm/common/membership.c
@@ -162,7 +162,7 @@ crm_node_t *crm_update_peer(
     node = g_hash_table_lookup(crm_peer_cache, uname);	
 
     if(node == NULL) {	
-	crm_debug("Creating entry for node %s/%u/%llu", uname, id, born);
+	crm_debug("Creating entry for node %s/%u/%llu", uname, id, (unsigned long long)born);
 	CRM_CHECK(id >= 0, return NULL);
 	CRM_CHECK(uuid != NULL, return NULL);
